requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
celery==5.4.0
redis==5.2.1
django-celery-results==2.5.1
django-storages==1.14.2
boto3==1.35.0